requirements/base.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
Django==4.2.19
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
language==0.6
numpy==2.0.2
openai==1.63.2
packaging==24.2
payme-pkg==3.0.24
pillow==11.1.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pylance==0.23.0
python-environ==0.4.54
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
sniffio==1.3.1
SpeechRecognition==3.14.2
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
